+++ CHANGELOG.md
@@ -1,8 +1,19 @@
 # Changelog
 
+## 1.0.0 🎉
+
+- Released on conda-forge
+- GDAL raster support by default in conda, optional in pip install
+- Removed Rasterio dependency, instead using GDAL directly
+- Added data manipulation methods FeatureClass to help with processing
+- Pass FeatureClass and FeatureDataset sequences on instantiation
+- Getter methods return dicts instead of tuples
+- Provided more details with get_info()
+- Dev environment setup/test/build batch scripts
+- Added TODO.md list for long-term goals
+
 ## 1.0.0-beta7
 
-- conda-forge distribution set up
 - Pass kwargs to the GeoTIFF write operation
 - Added get_info() and list_rasters() utility functions
 - Docs additions and updates